MediaProvider: Fix failure in android.provider.cts.MediaStore_FilesTest
openFileDescriptor or openOutputStream will call openFile in MediaProvider.
And the parent dir of the caller's url is created by openFile.
But MediaStore_FilesTest expect that the dir does not exists.
So the test case fail.

Remove the function of creating the dir to make MediaStore_FilesTest
successfully in openFile. And create the dir in the request function.

Change-Id: Id497e3ed70f41ca0a3a83a496ff82a5b3d33ca77
CRs-Fixed: 1020200
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>